requirements.txt
# =========================
# Core Python Utilities
# =========================
python-dotenv>=1.0.1
pydantic>=2.6
typing-extensions>=4.9

# =========================
# MCP (Model Context Protocol)
# =========================
mcp>=1.2.0
mcp[client]>=1.2.0

# =========================
# Agents & Orchestration
# =========================
smolagents>=0.4.0
langchain>=0.2.14
langchain-community>=0.2.12
langgraph>=0.2.22

# =========================
# LLM & Inference (NO OpenAI)
# =========================
litellm>=1.45.0
groq>=0.9.0

# =========================
# Evaluation
# =========================
deepeval>=3.8.0
pytest>=8.3.0
pytest-asyncio>=0.23.8

# =========================
# RAG / Embeddings (optional but aligned)
# =========================
sentence-transformers>=3.0.1
faiss-cpu>=1.8.0

# =========================
# Data / Parsing
# =========================
regex>=2024.5.15
rich>=13.7.1
